Restructure project into a hierarchy
