+++ arc-script/core/build.rs
@@ -1,5 +1,9 @@
 extern crate lalrpop;
 
 fn main() {
-    lalrpop::process_root().unwrap();
+    lalrpop::Configuration::new()
+        .emit_whitespace(false)
+        .use_cargo_dir_conventions()
+        .process()
+        .unwrap();
 }
