+++ arc-script/core/examples/emit.rs
@@ -1,9 +1,9 @@
-use arc_script::opt::*;
+use arc_script::prelude::*;
 fn main() {
     let mut args = std::env::args();
     let _ = args.next();
     let source = &args.next().unwrap();
     let opt = Opt::default();
-    let script = arc_script::compile(source, &opt);
+    let script = compiler::compile(source, &opt);
     println!("{}", script.emit_as_str());
 }
