+++ arc-script/core/src/cli/completer.rs
