+++ arc-script/core/src/cli/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "lsp")]
+pub mod completer;
+#[cfg(feature = "lsp")]
+pub mod linter;
+#[cfg(feature = "lsp")]
+pub mod lsp;
+#[cfg(feature = "repl")]
+pub mod repl;
