+++ arc-script/core/src/main.rs
@@ -1,15 +1,15 @@
 use anyhow::Result;
-use arc_script::{diagnose, io::*, opt::*};
+use arc_script::prelude::*;
 
 pub fn main() -> Result<()> {
     let opt = &Opt::parse();
     match &opt.subcmd {
-        SubCmd::File(c) => diagnose(&read_file(&c.path)?, opt),
-        SubCmd::Code(c) => diagnose(&c.code, opt),
+        SubCmd::File(c) => compiler::diagnose(&read_file(&c.path)?, opt),
+        SubCmd::Code(c) => compiler::diagnose(&c.code, opt),
         #[cfg(feature = "lsp")]
-        SubCmd::Lsp => arc_script::lsp::lsp(opt),
+        SubCmd::Lsp => lsp::start(opt),
         #[cfg(feature = "repl")]
-        SubCmd::Repl => arc_script::repl::repl(opt)?,
+        SubCmd::Repl => repl::start(opt)?,
         SubCmd::Lib => unreachable!(),
     }
     Ok(())
