+++ arc-script/core/src/passes/eval.rs
@@ -1,10 +1,5 @@
-use crate::{ast::*, dataflow::*, typer::*};
-use derive_more::Constructor;
+use crate::prelude::*;
 use std::collections::HashMap;
-use BinOpKind::*;
-use ExprKind::*;
-use LitKind::*;
-use UnOpKind::*;
 use Value::*;
 
 #[allow(unused)]
