+++ arc-script/core/src/passes/pruner.rs
