+++ arc-script/core/src/passes/shaper.rs
