+++ arc-script/core/src/passes/ssa.rs
@@ -1,4 +1,4 @@
-use crate::{info::*, prelude::*, utils::*};
+use crate::prelude::*;
 
 type Context = Vec<(Ident, Expr)>;
