+++ arc-script/core/src/passes/typer.rs
@@ -1,5 +1,5 @@
 use {
-    crate::{error::*, info::*, prelude::*, symbols::*},
+    crate::prelude::*,
     ena::unify::{InPlace, NoError, UnifyKey, UnifyValue},
 };
