+++ arc-script/core/src/prelude.rs
@@ -1,20 +1,29 @@
-pub use crate::ast::BinOpKind::*;
-pub use crate::ast::DeclKind::*;
-pub use crate::ast::DimKind::*;
-pub use crate::ast::DimOpKind::*;
-pub use crate::ast::ExprKind::*;
-pub use crate::ast::LitKind::*;
-pub use crate::ast::PatKind::*;
-pub use crate::ast::ScalarKind::*;
-pub use crate::ast::SettingKind::*;
-pub use crate::ast::TypeKind::*;
-pub use crate::ast::UnOpKind::*;
-pub use crate::ast::*;
-pub use crate::codegen::pretty::*;
-pub use crate::error::CompilerError;
-pub use crate::info::Info;
-pub use crate::opt::*;
-pub use crate::symbols::SymbolTable;
-pub use crate::typer::Typer;
-pub use codespan::Span;
-pub use derive_more::Constructor;
+pub(crate) use crate::codegen::pretty::*;
+pub(crate) use crate::passes::typer::Typer;
+pub(crate) use crate::passes::utils::*;
+pub(crate) use crate::repr::ast::BinOpKind::*;
+pub(crate) use crate::repr::ast::DeclKind::*;
+pub(crate) use crate::repr::ast::DimKind::*;
+pub(crate) use crate::repr::ast::DimOpKind::*;
+pub(crate) use crate::repr::ast::ExprKind::*;
+pub(crate) use crate::repr::ast::LitKind::*;
+pub(crate) use crate::repr::ast::PatKind::*;
+pub(crate) use crate::repr::ast::ScalarKind::*;
+pub(crate) use crate::repr::ast::SettingKind::*;
+pub(crate) use crate::repr::ast::TypeKind::*;
+pub(crate) use crate::repr::ast::UnOpKind::*;
+pub(crate) use crate::repr::ast::*;
+pub(crate) use crate::repr::dataflow::*;
+pub(crate) use crate::repr::info::error::{CompilerError, CompilerResult};
+pub(crate) use crate::repr::info::symbols::SymbolTable;
+pub(crate) use crate::repr::info::Info;
+pub(crate) use codespan::Span;
+pub(crate) use derive_more::Constructor;
+
+pub use crate::compiler;
+pub use crate::repr::info::opt::*;
+pub use crate::passes::io::*;
+#[cfg(feature = "repl")]
+pub use crate::cli::repl;
+#[cfg(feature = "lsp")]
+pub use crate::cli::lsp;
