+++ arc-script/core/src/repr/ast.rs
@@ -1,5 +1,4 @@
 use crate::prelude::*;
-use crate::{info::Info, typer::Typer};
 use chrono::Duration;
 use flat_map::flat_map::FlatMap;
 use lasso::Spur;
