+++ arc-script/core/src/repr/info/error.rs
